refactor: Split large testUtils.ts file to resolve code complexity issue
- Split 531-line testUtils.ts into 4 focused modules:
  - testDataFactories.ts (93 lines): Mock data creation functions
  - testAssertions.ts (137 lines): Common assertion helpers
  - testScenarios.ts (230 lines): Test scenario setup functions
  - testMocks.ts (196 lines): Mock utilities and specialized helpers
- testUtils.ts now only re-exports from focused modules (8 lines)
- All files under 500 lines to meet code complexity standards
- Maintained 100% test coverage and all 236 tests passing
- Fixed ESLint issues and imports

Resolves code complexity violation while preserving functionality

Relates to #106
